Refactoring UI tests, not to have everything in a single class
