+++ Lombiq.HelpfulExtensions.Tests.UI/Helpers/WidgetHelpers.cs
@@ -1,0 +1,18 @@
+using Atata;
+using Lombiq.Tests.UI.Extensions;
+using Lombiq.Tests.UI.Services;
+using OpenQA.Selenium;
+using OpenQA.Selenium.Interactions;
+using static Lombiq.HelpfulExtensions.Tests.UI.Constants.XPathSelectors;
+
+namespace Lombiq.HelpfulExtensions.Tests.UI.Helpers;
+
+internal static class WidgetHelpers
+{
+    public static void AddWidgetToPageFlow(UITestContext context, string widget) =>
+        new Actions(context.Driver)
+            .Click(context.Get(By.XPath(AddWidgetButton)))
+            .Click(context.Get(By.XPath($"{WidgetList}/a[text()='{widget} Widget']").OfAnyVisibility()))
+            .Build()
+            .Perform();
+}
